Adding tests for ctftime based functionality
