+++ CTFd/themes/original/templates/errors/403.html
@@ -2,12 +2,13 @@
 
 {% block content %}
 
-<div class="row">
-	<h1 class="text-center">403</h1>
-	<h2 class="text-center">An authorization error has occured</h2>
-	<h2 class="text-center">Please try again</h2>
+<div class="container main-container">
+	<div class="row">
+		<h1 class="text-center">403</h1>
+		<h2 class="text-center">An authorization error has occured</h2>
+		<h2 class="text-center">Please try again</h2>
+	</div>
 </div>
-
 
 {% endblock %}
