+++ CTFd/themes/original/templates/errors/404.html
@@ -2,10 +2,12 @@
 
 {% block content %}
 
-<div class="row">
-	<h1 class="text-center">404</h1>
-	<h2 class="text-center">Whoops, looks like we can't find that.</h2>
-	<h2 class="text-center">Sorry about that</h2>
+<div class="container main-container">
+	<div class="row">
+		<h1 class="text-center">404</h1>
+		<h2 class="text-center">Whoops, looks like we can't find that.</h2>
+		<h2 class="text-center">Sorry about that</h2>
+	</div>
 </div>
