+++ CTFd/themes/original/templates/errors/500.html
@@ -2,9 +2,11 @@
 
 {% block content %}
 
-<div class="row">
-	<h1 class="text-center">500</h1>
-	<h2 class="text-center">An Internal Server Error has occured</h2>
+<div class="container main-container">
+	<div class="row">
+		<h1 class="text-center">500</h1>
+		<h2 class="text-center">An Internal Server Error has occured</h2>
+	</div>
 </div>
