+++ CTFd/themes/original/templates/errors/502.html
@@ -2,9 +2,11 @@
 
 {% block content %}
 
-<div class="row">
-	<h1 class="text-center">502</h1>
-	<h2 class="text-center">That action isn't allowed</h2>
+<div class="container main-container">
+	<div class="row">
+		<h1 class="text-center">502</h1>
+		<h2 class="text-center">That action isn't allowed</h2>
+	</div>
 </div>
