+++ tests/admin/test_admin_facing.py
@@ -1,5 +1,8 @@
 from tests.helpers import *
 from CTFd.models import Teams
+from CTFd.utils import get_config, set_config, override_template, sendmail, verify_email, ctf_started, ctf_ended
+from freezegun import freeze_time
+from mock import patch
 
 
 def test_admin_panel():
@@ -71,4 +74,32 @@ def test_admin_config():
         client = login_as_user(app, name="admin", password="password")
         r = client.get('/admin/config')
         assert r.status_code == 200
+    destroy_ctfd(app)
+
+
+def test_admins_can_access_challenges_before_ctftime():
+    '''Admins can see and solve challenges despite it being before ctftime'''
+    app = create_ctfd()
+    with app.app_context():
+        set_config('start', '1507089600') # Wednesday, October 4, 2017 12:00:00 AM GMT-04:00 DST
+        set_config('end', '1507262400') # Friday, October 6, 2017 12:00:00 AM GMT-04:00 DST
+        register_user(app)
+        chal = gen_challenge(app.db)
+        chal_id = chal.id
+        flag = gen_flag(app.db, chal=chal.id, flag=u'flag')
+
+        with freeze_time("2017-10-2"):
+            client = login_as_user(app, name='admin', password='password')
+            r = client.get('/chals')
+            assert r.status_code == 200
+
+            with client.session_transaction() as sess:
+                data = {
+                    "key": 'flag',
+                    "nonce": sess.get('nonce')
+                }
+            r = client.post('/chal/{}'.format(chal_id), data=data)
+            assert r.status_code == 200
+        solve_count = app.db.session.query(app.db.func.count(Solves.id)).first()[0]
+        assert solve_count == 1
     destroy_ctfd(app)
